\name{isolve-package}
\alias{isolve-package}
\alias{isolve}
\docType{package}
\title{
\packageTitle{isolve}
}
\description{
\packageDescription{isolve}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{isolve}
\packageIndices{isolve}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{isolve}

Maintainer: \packageMaintainer{isolve}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
